requirements.txt
mlflow
scikit-learn
fastapi
uvicorn
pandas
numpy
pyyaml
evidently
joblib
pydantic
